requirements.txt
# Multi-Tenant Schema Translator - Python Requirements

# FastAPI and web server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Azure OpenAI and AI
openai>=1.54.0
tiktoken==0.5.2

# Database connectivity
pyodbc==5.0.1
aioodbc==0.4.0
aiosqlite==0.19.0
# Note: On Azure App Service, ODBC drivers are pre-installed

# Environment and configuration
python-dotenv==1.0.0

# CORS and security
python-jose[cryptography]==3.3.0

# Data processing
pydantic==2.5.3
pydantic-settings==2.1.0
pyyaml==6.0.1

# Utilities
requests==2.31.0
aiofiles==23.2.1

# Error resilience and retries
tenacity==8.2.3

# For static file serving
jinja2==3.1.3
